requirements-tensorflow.txt
tensorflow~=2.5.0
openvino-dev==2021.4.1
